requirements.txt
msal==1.31.1
python-docx==1.1.2
requests==2.32.3
openai>=1.55.0
python-dotenv==1.0.1
loguru==0.7.2
